backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
supabase>=2.10.0
pydantic>=2.10.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
